Fix mergeObjects
For some reason `mergeObjects` breaks in our environment but works in superagent-mocker tests.

In our case, the `arguments` object does not iterate with `for(var in arguments)`. Additionally, `Object.keys(arguments)` returns an empty array. This is in line with what I would expect from the `arguments` object.

I haven't been able to figure out why this does not break in superagent-mocker tests, but this fix works in both situations.